data-engineering/bigquery-with-bacalhau/process_container/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile - -o requirements.txt
cachetools==5.5.0
    # via google-auth
certifi==2024.12.14
    # via requests
charset-normalizer==3.4.1
    # via requests
duckdb==1.1.3
faker==33.3.1
google-api-core==2.24.0
    # via
    #   google-cloud-bigquery
    #   google-cloud-core
    #   google-cloud-iam
    #   google-cloud-resource-manager
    #   google-cloud-service-usage
    #   google-cloud-storage
google-auth==2.37.0
    # via
    #   google-api-core
    #   google-cloud-bigquery
    #   google-cloud-core
    #   google-cloud-iam
    #   google-cloud-resource-manager
    #   google-cloud-service-usage
    #   google-cloud-storage
google-cloud-bigquery==3.27.0
google-cloud-core==2.4.1
    # via
    #   google-cloud-bigquery
    #   google-cloud-storage
google-cloud-iam==2.17.0
google-cloud-resource-manager==1.14.0
google-cloud-service-usage==1.12.0
google-cloud-storage==2.19.0
google-crc32c==1.6.0
    # via
    #   google-cloud-storage
    #   google-resumable-media
google-resumable-media==2.7.2
    # via
    #   google-cloud-bigquery
    #   google-cloud-storage
googleapis-common-protos==1.66.0
    # via
    #   google-api-core
    #   grpc-google-iam-v1
    #   grpcio-status
grpc-google-iam-v1==0.14.0
    # via
    #   google-cloud-iam
    #   google-cloud-resource-manager
grpcio==1.69.0
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
grpcio-status==1.69.0
    # via google-api-core
idna==3.10
    # via requests
natsort==8.4.0
numpy==2.2.1
    # via pandas
packaging==24.2
    # via google-cloud-bigquery
pandas==2.2.3
pip==24.3.1
proto-plus==1.25.0
    # via
    #   google-api-core
    #   google-cloud-iam
    #   google-cloud-resource-manager
    #   google-cloud-service-usage
protobuf==5.29.3
    # via
    #   google-api-core
    #   google-cloud-iam
    #   google-cloud-resource-manager
    #   google-cloud-service-usage
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
    #   proto-plus
pyarrow==18.1.0
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.1
    # via google-auth
python-dateutil==2.9.0.post0
    # via
    #   faker
    #   google-cloud-bigquery
    #   pandas
pytz==2024.2
    # via pandas
pyyaml==6.0.2
requests==2.32.3
    # via
    #   google-api-core
    #   google-cloud-bigquery
    #   google-cloud-storage
rsa==4.9
    # via google-auth
setuptools==75.8.0
six==1.17.0
    # via python-dateutil
typing-extensions==4.12.2
    # via faker
tzdata==2024.2
    # via pandas
urllib3==2.3.0
    # via requests
wheel==0.45.1
